refactor(asr): 提取 Base64Encode 为公共工具 (utils/base64)
