+++ src/addon/asr/openai_asr.cpp
@@ -12,6 +12,8 @@
 #include <fcitx-utils/log.h>
 #include <json/json.h>
 
+#include "utils/base64.h"
+
 using namespace std::chrono_literals;
 
 namespace fcitx {
@@ -60,25 +62,7 @@ size_t WriteCallback(void* contents, size_t size, size_t nmemb, void* userp) {
     return size * nmemb;
 }
 
-// Base64 encode
-static const char b64table[] =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/="
-;
-
-std::string Base64Encode(const uint8_t* data, size_t len) {
-    std::string out;
-    out.reserve(((len + 2) / 3) * 4);
-    for (size_t i = 0; i < len; i += 3) {
-        uint32_t v = (static_cast<uint32_t>(data[i]) << 16)
-                   | (i + 1 < len ? static_cast<uint32_t>(data[i + 1]) << 8 : 0)
-                   | (i + 2 < len ? static_cast<uint32_t>(data[i + 2]) : 0);
-        out.push_back(b64table[(v >> 18) & 0x3F]);
-        out.push_back(b64table[(v >> 12) & 0x3F]);
-        out.push_back(i + 1 < len ? b64table[(v >> 6) & 0x3F] : '=');
-        out.push_back(i + 2 < len ? b64table[v & 0x3F] : '=');
-    }
-    return out;
-}
+// Base64 编码已提取到 utils/base64.h，此处复用公共实现。
 
 std::string BuildMultipartBody(const std::vector<uint8_t>& wavData,
                                const std::string& model,
